Refactored stream reading
An H264 stream is read continuously spitting out NAL units as they appear. The NAL units are put in BitReaders. A NAL unit itself now correctly keeps track of the RBSP bytes.

Meanwhile, the H264 stream keeps track of different video streams, as denoted by the receipt of an SPS NAL.

All this is to support the peeking features of CABAC and thus ae, the context-adaptive arithmetic entropy coded syntax element.